pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-ollama>=0.3.1",
    "langchain-openai>=0.3.12",
    "langchain-qdrant>=0.2.0",
    "markdownify>=1.1.0",
    "pymupdf>=1.25.5",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.13.3",
    "sentence-transformers>=4.0.2",
    "streamlit>=1.44.1",
    "watchdog>=6.0.0",
]
